+++ source/Application/OutgoingMessages/DocumentFactory.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Domain.Documents;
 using Domain.OutgoingMessages;
 using NodaTime;
