+++ source/Infrastructure/IncomingMessages/CimFormatParser.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Linq;
+using Domain.Documents;
 using Domain.OutgoingMessages;
 using Domain.SeedWork;
