+++ source/Infrastructure/IncomingMessages/RequestChangeAccountingPointCharacteristics/XmlMessageParser.cs
@@ -24,7 +24,7 @@ using CimMessageAdapter.Messages;
 using DocumentValidation;
 using DocumentValidation.CimXml;
 using Domain.OutgoingMessages;
-using DocumentFormat = Domain.OutgoingMessages.DocumentFormat;
+using DocumentFormat = Domain.Documents.DocumentFormat;
 using MarketActivityRecord = Application.IncomingMessages.RequestChangeAccountPointCharacteristics.MarketActivityRecord;
 
 namespace Infrastructure.IncomingMessages.RequestChangeAccountingPointCharacteristics;
